General packaging updates and refactoring
Updated Python packaging style to conform with newer standards,
refactored general repo layout.
