--- brainprint/__init__.py
@@ -1,1 +1,0 @@
-name = "brainprint"
