+++ pyproject.toml
@@ -1,0 +1,30 @@
+[tool.black]
+exclude = '''
+# Directories
+/(
+    \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | venv
+  | _build
+  | build
+  | dist
+)/
+'''
+line-length = 79
+
+[tool.coverage.run]
+source = ["brainprint"]
+branch = true
+omit = ["*tests*", "venv/**", "setup.py"]
+command_line = "-m pytest tests"
+
+[tool.coverage.paths]
+source = ["src"]
+
+[build-system]
+requires = ["setuptools>=42", "wheel"]
+build-backend = "setuptools.build_meta"
