+++ setup.cfg
@@ -1,0 +1,63 @@
+[metadata]
+name = brainprint
+version = 0.2.0
+author = Kersten Diers, Martin Reuter
+author_email = kersten.diers@dzne.de
+description = A light-weight package derived from the BrainPrint scripts
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = MIT
+license_files = LICENSE
+keywords = neuroscience, sMRI, FreeSurfer
+url = https://github.com/Deep-MI/BrainPrint-python
+project_urls =
+    Source = https://github.com/Deep-MI/BrainPrint-python
+    Bug Reports = https://github.com/Deep-MI/BrainPrint-python/issues/
+classifiers =
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3 :: Only
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6, <4
+include_package_data = True
+install_requires =
+    lapy @ git+https://github.com/Deep-MI/LaPy.git#egg=lapy
+    numpy
+    pandas
+entry_points =
+    [console_scripts]
+    brainprint = brainprint.brainprint:run_brainprint
+
+[options.extras_require]
+dev =
+    black
+    flake8
+    ipython
+
+[options.packages.find]
+where = src
+
+
+
+[flake8]
+max-complexity = 10
+exclude =
+    __pycache__,
+    .git,
+    .pytest_cache,
+    .tox,
+    build,
+    dist,
+    tests,
+    venv
+per-file-ignores =
+    */__init__.py: F401
+    */messages.py: E501
