+++ src/brainprint/__init__.py
@@ -1,0 +1,3 @@
+"""
+BrainPrint: A script to compute ShapeDNA of FreeSurfer structures.
+"""
