+++ src/brainprint/brainprint.py
@@ -1,428 +1,22 @@
 """
-brainprint - a script to compute ShapeDNA of FreeSurfer structures
-
-type 'brainprint.get_help()' for info.
+Script entry point.
 """
-
-# ==============================================================================
-# FUNCTIONS
-
-# ------------------------------------------------------------------------------
-# help function
-
-
-def get_help(print_help=True):
-    """
-    a function to retrieve a help message
-    """
-
-    HELPTEXT = """
-
-    brainprint.py
-    Author: Martin Reuter, 2015
-
-    =======
-    SUMMARY
-    =======
-
-    Computes the BrainPrint for a FreeSurfer subject.
-
-    The BrainPrint consists of the shape descriptors (Shape-DNA) [1]
-    of a selection of both cortical and subcortical structures [2].
-
-    Here is a list of structures and FreeSurfer aseg label ids:
-
-    CorpusCallosum                  [251, 252, 253, 254, 255]
-    Cerebellum                      [7, 8, 16, 46, 47]
-    Ventricles                      [4, 5, 14, 24, 31, 43, 44, 63]
-    3rd-Ventricle                   [14, 24]
-    4th-Ventricle                   15
-    Brain-Stem                      16
-    Left-Striatum                   [11, 12, 26]
-    Left-Lateral-Ventricle          [4, 5, 31]
-    Left-Cerebellum-White-Matter    7
-    Left-Cerebellum-Cortex          8
-    Left-Thalamus-Proper            10
-    Left-Caudate                    11
-    Left-Putamen                    12
-    Left-Pallidum                   13
-    Left-Hippocampus                17
-    Left-Amygdala                   18
-    Left-Accumbens-area             26
-    Left-VentralDC                  28
-    Right-Striatum                  [50, 51, 58]
-    Right-Lateral-Ventricle         [43, 44, 63]
-    Right-Cerebellum-White-Matter   46
-    Right-Cerebellum-Cortex         47
-    Right-Thalamus-Proper           49
-    Right-Caudate                   50
-    Right-Putamen                   51
-    Right-Pallidum                  52
-    Right-Hippocampus               53
-    Right-Amygdala                  54
-    Right-Accumbens-area            58
-    Right-VentralDC                 60
-
-    And the following cortical structures:
-
-    lh-white-2d    (left white matter surface triangles)
-    lh-pial-2d     (left pial surface triangles)
-    rh-white-2d    (same for right hemisphere ...)
-    rh-pial-2d
-
-    Processing of the cortical structures can be skipped (--skipcortex).
-
-    Implicit Inputs:
-    The mri/aseg.mgz and mri/norm.mgz should be available.
-    Also surf/?h.pial and surf/?h.white need to be
-    available unless --skipcortex is passed. norm.mgz is not
-    absolutely necessary but highly recommended to fix the labels
-    and obtain improved meshes.
-
-    Output:
-    The brainprint CSV table containing column headers for the
-    structures, a row of areas, a row of volumes and N rows of
-    the first N eigenvalues for each structure. An additional
-    output file will be created if the asymmetry calculation is
-    performed.
-
-    ==================
-    COMMAND-LINE USAGE
-    ==================
-
-    python3 brainprint.py --sdir <directory> --sid <SubjectID>  [--num <num>]
-                        [--evec] [--skipcortex] [--norm <surface|volume|geometry|none> ]
-                        [--reweight] [--asymmetry] [--outdir <directory>] [--help]
-                        [--more-help]
-
-    Options:
-      --help           Show this help message and exit
-      --more-help      Show extensive help message and exit
-
-    Required options:
-      --sid <SubjectID>
-                       Subject ID (FreeSurfer-processed directory inside the
-                       subjects directory)
-      --sdir <directory>
-                       FreeSurfer subjects directory
-
-    Processing directives:
-      --num <num>      Number of eigenvalues/vectors to compute (default: 50)
-      --evec           Switch on eigenvector computation (default: off)
-      --skipcortex     Skip cortical surfaces (default: off)
-      --norm <surface|volume|geometry|none>
-                       Switch on eigenvalue normalization; will be either surface,
-                       volume, or determined by the geometry of the object. Use
-                       "none" or leave out entirely to skip normalization.
-      --reweight       Switch on eigenvalue reweighting (default: off)
-      --asymmetry      Perform left-right asymmetry calculation (default: off)
-
-    Output parameters:
-      --outdir=OUTDIR  Output directory (default: <sdir>/<sid>/brainprint)
-
-    ============
-    PYTHON USAGE
-    ============
-
-    As an alternative to the command-line usage described above, individual
-    functions can also be called within a Python environment as follows:
-
-    import lapy
-    from brainprint import brainprint
-    brainprint.run_brainprint(sdir="/my/subjects/directory", sid="my_subject_id")
-
-    Additional options are num=<int>, evec=<bool>, skipcortex=<bool>,
-    norm=<"surface"|"volume"|"geometry"|"none">, reweight=<bool>, and outdir=<string>.
-
-    See `help(brainprint)` and `brainprint.get_help()` for further usage info
-    and additional options.
-
-    =============
-    REQUIREMENTS
-    ============
-
-    The script requires the lapy package, which can be installed from
-    https://github.com/Deep-MI/LaPy using
-    pip3 install --user git+https://github.com/Deep-MI/LaPy.git.
-
-    ==========
-    REFERENCES
-    ==========
-
-    If used for a publication, please cite both [1] for the shape
-    descriptor method and [2] for the application to brain MRI and
-    definiton of the BrainPrint.
-
-    [1] M. Reuter, F.-E. Wolter and N. Peinecke.
-    Laplace-Beltrami spectra as "Shape-DNA" of surfaces and solids.
-    Computer-Aided Design 38 (4), pp.342-366, 2006.
-    http://dx.doi.org/10.1016/j.cad.2005.10.011
-
-    [2] C. Wachinger, P. Golland, W. Kremen, B. Fischl, M. Reuter.
-    BrainPrint: A discriminative characterization of brain morphology.
-    NeuroImage Volume 109, pp.232-248, 2015.
-    http://dx.doi.org/10.1016/j.neuroimage.2015.01.032
-
-    """
-
-    if print_help is True:
-        print(HELPTEXT)
-    else:
-        return HELPTEXT
-
-
-# ------------------------------------------------------------------------------
-# parse options
-
-# parse_options
-def _parse_options():
-    """
-    command line options parser: initiate the option parser and return the parsed object
-    """
-
-    # imports
-    import argparse
-    import sys
-
-    # setup parser
-    parser = argparse.ArgumentParser(
-        description="This program conducts a brainprint analysis of FreeSurfer output.",
-        add_help=False,
-    )
-
-    # help text
-    h_sid = "Subject ID (FreeSurfer processed directory inside the subjects directory)"
-    h_sdir = "FreeSurfer subjects directory"
-    h_outdir = "Output directory (default: <sdir>/<sid>/brainprint)"
-    h_num = "Number of eigenvalues/vectors to compute (default: 50)"
-    h_evec = "Switch on eigenvector computation (default: off)"
-    h_skipcortex = "Skip cortical surfaces (default: off)"
-    h_norm = "Perform on eigenvalue normalization (default: none)"
-    h_rwt = "Switch on eigenvalue reweighting (default: off)"
-    h_asym = "Switch on additional asymmetry calculation (default: off)"
-
-    # required arguments
-    required = parser.add_argument_group("Required arguments")
-
-    required.add_argument(
-        "--sid",
-        dest="sid",
-        help=h_sid,
-        default=None,
-        metavar="<string>",
-        required=False,
-    )
-    required.add_argument(
-        "--sdir",
-        dest="sdir",
-        help=h_sdir,
-        default=None,
-        metavar="<directory>",
-        required=False,
-    )
-
-    # optional arguments
-    optional = parser.add_argument_group("Processing directives")
-
-    optional.add_argument(
-        "--num",
-        dest="num",
-        help=h_num,
-        default=50,
-        metavar="<num>",
-        type=int,
-        required=False,
-    )
-    optional.add_argument(
-        "--evec",
-        dest="evec",
-        help=h_evec,
-        default=False,
-        action="store_true",
-        required=False,
-    )
-    optional.add_argument(
-        "--skipcortex",
-        dest="skipcortex",
-        help=h_skipcortex,
-        default=False,
-        action="store_true",
-        required=False,
-    )
-    optional.add_argument(
-        "--norm",
-        dest="norm",
-        help=h_norm,
-        default="none",
-        metavar=" <surface|volume|geometry|none>",
-        required=False,
-    )
-    optional.add_argument(
-        "--reweight",
-        dest="rwt",
-        help=h_rwt,
-        default=False,
-        action="store_true",
-        required=False,
-    )
-    optional.add_argument(
-        "--asymmetry",
-        dest="asymmetry",
-        help=h_asym,
-        default=False,
-        action="store_true",
-        required=False,
-    )
-
-    # output options
-    output = parser.add_argument_group("Output parameters")
-
-    output.add_argument(
-        "--outdir",
-        dest="outdir",
-        help=h_outdir,
-        default=None,
-        metavar="<directory>",
-        required=False,
-    )
-
-    # define help
-    help = parser.add_argument_group("Getting help")
-
-    help.add_argument(
-        "--help", help="Display this help message and exit", action="help"
-    )
-    help.add_argument(
-        "--more-help",
-        dest="more_help",
-        help="Display extensive help message and exit",
-        default=False,
-        action="store_true",
-        required=False,
-    )
-
-    # --------------------------------------------------------------------------
-    #
-
-    # check if there are any inputs; if not, print help and exit
-    if len(sys.argv) == 1:
-        args = parser.parse_args(["--help"])
-    else:
-        args = parser.parse_args()
-
-    # return extensive helptext
-    if args.more_help is True:
-        get_help()
-        sys.exit(0)
-
-    # check for required arguments print help and exit
-    if args.sid is None:
-        print(
-            "ERROR: the --sid argument is required, exiting. Use --help to see details."
-        )
-        sys.exit(1)
-
-    if args.sdir is None:
-        print(
-            "ERROR: the --sdir argument is required, exiting. Use --help to see details."
-        )
-        sys.exit(1)
-
-    # convert options to dict
-    options = dict(
-        sdir=args.sdir,
-        sid=args.sid,
-        outdir=args.outdir,
-        num=args.num,
-        evec=args.evec,
-        skipcortex=args.skipcortex,
-        norm=args.norm,
-        rwt=args.rwt,
-        asymmetry=args.asymmetry,
-    )
-
-    # return
-    return options
-
-
-# check_options
-def _check_options(options):
-    """
-    a function to evaluate input options and set some defaults
-    """
-
-    # imports
-    import errno
-    import os
-    import sys
-
-    # check if there are any inputs
-    if options["sdir"] is None and options["sid"] is None:
-        get_help(print_help=True)
-        sys.exit(0)
-
-    #
-    if options["sdir"] is None:
-        print("\nERROR: specify subjects directory via --sdir\n")
-        sys.exit(1)
-
-    if options["sid"] is None:
-        print("\nERROR: Specify --sid\n")
-        sys.exit(1)
-
-    subjdir = os.path.join(options["sdir"], options["sid"])
-    if not os.path.exists(subjdir):
-        print("\nERROR: cannot find sid in subjects directory\n")
-        sys.exit(1)
-
-    if options["outdir"] is None:
-        options["outdir"] = os.path.join(subjdir, "brainprint")
-    try:
-        os.mkdir(options["outdir"])
-        os.mkdir(os.path.join(options["outdir"], "eigenvectors"))
-        os.mkdir(os.path.join(options["outdir"], "surfaces"))
-        os.mkdir(os.path.join(options["outdir"], "temp"))
-    except OSError as e:
-        if e.errno != errno.EEXIST:
-            raise e
-        pass
-
-    # return
-    return options
-
-
-# ------------------------------------------------------------------------------
-# auxiliary functions
-
-
-def _run_cmd(cmd, err_msg, expected_retcode=[0]):
-    """
-    execute the command
-    """
-
-    # imports
-    import shlex
-    import subprocess
-    import sys
-
-    # run cmd
-    print("#@# Command: " + cmd + "\n")
-    args = shlex.split(cmd)
-    retcode = subprocess.call(args)
-    if (retcode in expected_retcode) is False:
-        print("ERROR: " + err_msg)
-        sys.exit(1)
+import importlib
+import os
+import subprocess
+import sys
+import uuid
+import warnings
+
+from cli.parser import _check_options, _parse_options, _run_cmd
+
+warnings.filterwarnings("ignore", ".*negative int.*")
 
 
 def _get_ev(evfile):
     """
     returns string list of area, volume and evals
     """
-
-    # imports
-    import os
-
-    #
     if not os.path.isfile(evfile):
         return
     area = ""
@@ -438,7 +32,7 @@ def _get_ev(evfile):
             if "Eigenvalues:" in line:
                 evline = next(inF)
                 evstr = ""
-                while (evline is not None) and (not "}" in evline):
+                while (evline is not None) and ("}" not in evline):
                     evstr = evstr + evline
                     evline = next(inF)
                 evstr = evstr + evline
@@ -460,10 +54,6 @@ def _write_ev(options, evMat, evecMat=None, dstMat=None):
     """
     writes EV files
     """
-
-    # imports
-    import os
-
     import numpy as np
     import pandas as pd
 
@@ -510,12 +100,6 @@ def _get_aseg_surf(options):
     """
     a function to create a surface from the aseg and label files
     """
-
-    # imports
-    import os
-    import uuid
-
-    #
     astring = " ".join(options["asegid"])
     subjdir = os.path.join(options["sdir"], options["sid"])
     aseg = os.path.join(subjdir, "mri", "aseg.mgz")
@@ -560,11 +144,6 @@ def _compute_brainprint(options):
     """
     a function to compute shapeDNA descriptors for several structures
     """
-
-    # imports
-    import os
-    import subprocess
-
     import numpy as np
     from lapy import ShapeDNA, TriaIO, TriaMesh, read_geometry
 
@@ -574,7 +153,8 @@ def _compute_brainprint(options):
     # - Right Striatum: right Caudate + Putamen + Accumbens
     # - CorpusCallosum: 5 subregions combined
     # - Cerebellum: brainstem + (left+right) cerebellum WM and GM
-    # - Ventricles: (left+right) lat.vent + inf.lat.vent + choroidplexus + 3rdVent + CSF
+    # - Ventricles: (left+right) lat.vent + inf.lat.vent + choroidplexus
+    # + 3rdVent + CSF
     # - Lateral-Ventricle: lat.vent + inf.lat.vent + choroidplexus
     # - 3rd-Ventricle: 3rd-Ventricle + CSF
 
@@ -655,7 +235,7 @@ def _compute_brainprint(options):
 
             # message
             print(
-                "\n\n==========================================================="
+                "\n\n========================================================="
             )
             print(
                 "2D Cortical Surface "
@@ -742,7 +322,7 @@ def _compute_brainprint(options):
                 )
             )
 
-        except subprocess.CalledProcessError as e:
+        except subprocess.CalledProcessError:
             print("Error occured, skipping label " + surfaces_i)
             failed = True
 
@@ -761,12 +341,8 @@ def _compute_asymmetry(options, evMat):
     a function to compute lateral shape analysis using the brainprint
     """
 
-    # imports
-    import os
-    import subprocess
-
     import numpy as np
-    from lapy import ShapeDNA, TriaIO
+    from lapy import ShapeDNA
 
     # define structures
     # combined and individual aseg labels:
@@ -774,7 +350,8 @@ def _compute_asymmetry(options, evMat):
     # - Right Striatum: right Caudate + Putamen + Accumbens
     # - CorpusCallosum: 5 subregions combined
     # - Cerebellum: brainstem + (left+right) cerebellum WM and GM
-    # - Ventricles: (left+right) lat.vent + inf.lat.vent + choroidplexus + 3rdVent + CSF
+    # - Ventricles: (left+right) lat.vent + inf.lat.vent + choroidplexus
+    # + 3rdVent + CSF
     # - Lateral-Ventricle: lat.vent + inf.lat.vent + choroidplexus
     # - 3rd-Ventricle: 3rd-Ventricle + CSF
 
@@ -860,11 +437,10 @@ def run_brainprint(
     """
     a function to run a BrainPrint analysis
     """
+    from brainprint import messages
 
-    # imports
-    import importlib
-    import os
-    import sys
+    if options is None:
+        options = _parse_options()
 
     # get options
     if options is None:
@@ -898,7 +474,7 @@ def run_brainprint(
     try:
         cmd = "mri_binarize -version"
         _run_cmd(cmd, "mri_binarize failed.", expected_retcode=[0])
-    except:
+    except Exception:
         print("ERROR: could not find / run FreeSurfer binaries.")
         sys.exit(1)
 
@@ -925,27 +501,5 @@ def run_brainprint(
     # write EVs
     _write_ev(options, evMat, evecMat, dstMat)
 
-    # return
-    print(
-        "Returning matrices for eigenvalues, eigenvectors, and (optionally) distances."
-    )
-    print("The eigenvalue matrix contains area and volume as first two rows.")
+    print(messages.RETURN_VALUES)
     return evMat, evecMat, dstMat
-
-
-# ------------------------------------------------------------------------------
-# main function
-
-if __name__ == "__main__":
-
-    # imports
-    import warnings
-
-    # settings
-    warnings.filterwarnings("ignore", ".*negative int.*")
-
-    # parse command line options
-    options = _parse_options()
-
-    # run brainprint
-    run_brainprint(options)
