+++ src/brainprint/messages.py
@@ -1,0 +1,5 @@
+"""
+String messages for the :mod:`brainprint` package.
+"""
+RETURN_VALUES: str = """Returning matrices for eigenvalues, eigenvectors, and (optionally) distances.
+The eigenvalue matrix contains area and volume as first two rows."""
