+++ src/cli/__init__.py
@@ -1,0 +1,3 @@
+"""
+BrainPrint analysis CLI.
+"""
