+++ src/cli/utils.py
@@ -1,0 +1,14 @@
+"""
+Utility functions for the :mod:`brainprint.cli` module.
+"""
+from cli import help_text
+
+
+def get_help(print_help: bool = True):
+    """
+    Returns a detailed help message.
+    """
+    if print_help:
+        print(help_text.HELPTEXT)
+    else:
+        return help_text.HELPTEXT
